security: lock down test-session + verify installation ownership
Two real vulnerabilities found in the audit, both fixable in <50 LOC.

### CRITICAL — /api/auth/test-session prod takeover
The endpoint was gated only by ENABLE_TEST_SESSION === 'true'. If that
env var is ever set in the prod project (one Vercel-dashboard typo away)
an attacker can POST { username: 'lantos1618' } and receive a signed
better-auth.session_token cookie for that user — Pro tier, no GitHub
OAuth needed.

Fixed with defense in depth: hard-block when NODE_ENV === 'production'
AND VERCEL_ENV === 'production', regardless of ENABLE_TEST_SESSION. The
flag continues to gate dev and preview environments as before, and
e2e tests still work — they run against a non-production VERCEL_ENV.

### HIGH — /api/auth/link-installation hijack
Any signed-in user could POST { installationId: X } to claim any
unlinked GitHub App installation — granting them access to that org's
private repos via our app token. The endpoint only checked "is the
installation already linked to someone else?" — never "does the caller
actually have access to it?"

Installation IDs aren't secret (they appear in webhook deliveries,
post-install redirect URLs, and are incrementing integers).

Fixed: call GET /user/installations with the caller's OAuth token
before linking. GitHub returns only installations accessible to the
authenticated user (org member or installer), so a presence check
there is sufficient. 403 if absent.

Also tightened input validation (Number.isFinite, positive int) and
removed log lines that leaked installationId + userId into traces.

refactor: split profile router god file into 5 sub-routers

src/lib/trpc/routes/profile.ts was a 1019-LOC, 12-procedure god file
covering reads, generation, search, email lookup, and repo listings.

Split following the existing github/ subdirectory pattern:

  profile/
  ├── index.ts       — composes sub-routers
  ├── repos.ts       — getUserRepositories
  ├── email.ts       — getDeveloperEmail
  ├── read.ts        — generateProfile (cached), publicGetProfile,
  │                    getProfileVersions, getProfileByVersion,
  │                    getAnalyzedProfileCount, getAllAnalyzedProfiles
  ├── generate.ts    — checkGenerationStatus, generateProfileMutation,
  │                    clearProfileCache
  └── search.ts      — searchProfiles (semantic + filter)

Largest file (generate.ts) is 449 LOC, driven by the necessarily-long
streaming subscription. Total down from 1019 → 881 LOC after removing
~140 LOC of incidental verbose console logging and dead variables.

No public API change — same procedures at the same paths.
